requirements-test.txt
pytest==8.2.2
pytest-flask==1.3.0
pylint==2.12.2
safety==2.3.5
bandit==1.7.4
